Refactor library for Vue v3 support
Wherever possible, I have kept as close to the Vue v2 code as I could,
but I was forced to make typing judgement calls in a few places that
should probably be double-checked.
The only exception to the hands-off approach is in the example app,
which I took the liberty of porting to the Vue v3 composition API since
I'd be touching most of the LOC in it anyway.