Enable tree-shaking (really it was this easy?!)
Note: this tree-shakes, but only mostly due to Webpack 4 struggling with
"double-barrel" structures like we have with `index.ts` and `entry.ts`.
Importing a single icon from Phosphor-Vue is not a full 2.5MB anymore,
but there's still a bunch of unnecessary code included (it *does* seem
to purge the unused templates which is the bulk of the library size).
Webpack 5 apparently handles double-barrels better, so we should see
that problem disappear once vue-cli ships Webpack 5 support and clients
start using it (track https://github.com/vuejs/vue-cli/pull/5611)